deleting image - back-end

+++ springBoot/src/main/java/io/github/samirsales/Service/PictureService.java
@@ -4,9 +4,14 @@ import io.github.samirsales.Dao.PictureDao;
 import io.github.samirsales.Entity.Picture;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.nio.file.Files;
 import java.util.List;
 
 @Service
@@ -15,6 +20,9 @@ public class PictureService {
     @Autowired
     @Qualifier("fakeData")
     private PictureDao pictureDao;
+
+    @Value("${uploading.image.path}")
+    private String imagePath;
 
     public Picture getPictureById(Long id) {
         return pictureDao.getPictureById(id);
@@ -28,19 +36,43 @@ public class PictureService {
         return pictureDao.getPicturesBySearch(search);
     }
 
-    public Picture insertPicture(Picture picture) {
+    public Picture insertPicture(MultipartFile file) throws IOException {
+
+        Picture picture = pictureDao.insertPicture(new Picture());
+        picture.setFileName(picture.getId()+"."+getFileExtension(file.getOriginalFilename()));
+
+        String path = imagePath+"/"+picture.getFileName();
+        File convertFile = new File(path);
+        System.out.println(path);
+
+        try {
+            convertFile.createNewFile();
+            FileOutputStream fout = new FileOutputStream(convertFile);
+            fout.write(file.getBytes());
+            fout.close();
+        } catch (IOException ioException) {
+            ioException.printStackTrace();
+            removePictureById(picture.getId());
+            throw ioException;
+        }
+
         return pictureDao.insertPicture(picture);
     }
 
-    public void removePictureById(long id) {
-        pictureDao.removePictureById(id);
+    public void removePictureById(long id) throws IOException {
+        Picture picture = pictureDao.getPictureById(id);
+        if(picture != null){
+            File file = new File(imagePath+"/"+picture.getFileName());
+            Files.deleteIfExists(file.toPath());
+            pictureDao.removePictureById(id);
+        }
     }
 
     public void updatePicture(Picture picture) {
         pictureDao.updatePicture(picture);
     }
 
-    public String getFileExtension(String fullName) {
+    private String getFileExtension(String fullName) {
         if(fullName.isEmpty()){
             return "";
         }
